requirements.txt
SQLAlchemy
selenium
PyYAML
tornado
